Create schema.py to expose a single public module to CPython and Pypy
