+++ fastavro/_schema_py.py
@@ -3,33 +3,7 @@
 from os import path
 import json
 
-PRIMITIVES = set([
-    'boolean',
-    'bytes',
-    'double',
-    'float',
-    'int',
-    'long',
-    'null',
-    'string',
-])
-
-SCHEMA_DEFS = {
-    'boolean': 'boolean',
-    'bytes': 'bytes',
-    'double': 'double',
-    'float': 'float',
-    'int': 'int',
-    'long': 'long',
-    'null': 'null',
-    'string': 'string',
-}
-
-
-class UnknownType(Exception):
-    def __init__(self, name):
-        super(UnknownType, self).__init__(name)
-        self.name = name
+from ._schema_public import PRIMITIVES, SCHEMA_DEFS, UnknownType
 
 
 def extract_record_type(schema):
